LayoutTests/fast/scrolling/scroll-animator-basic-events-expected.txt

CONSOLE MESSAGE: line 24: MainFrameView: didAddVerticalScrollbar
CONSOLE MESSAGE: line 24: RenderLayer: didAddVerticalScrollbar
CONSOLE MESSAGE: line 24: MainFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 24: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 25: RenderLayer: mouseEnteredContentArea
CONSOLE MESSAGE: line 25: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 25: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 26: RenderLayer: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: line 26: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: line 26: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 27: RenderLayer: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: line 28: RenderLayer: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: line 29: RenderLayer: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: line 29: RenderLayer: mouseExitedContentArea
CONSOLE MESSAGE: line 29: MainFrameView: mouseMovedInContentArea
PASS successfullyParsed is true

TEST COMPLETE
This is a scrollable div.
